requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
requests==2.31.0
boto3==1.34.144
gunicorn==21.2.0
python-dotenv==1.0.0
PyMySQL==1.1.0
Pillow==10.3.0
cryptography==42.0.5
